Rename pages to screens, SgnedInOut to SignInOut
