+++ src/screens/StudyScreen/AlphabetSlider.jsx
