+++ src/screens/StudyScreen/FinishStudy.jsx
