+++ src/screens/StudyScreen/WordCard.jsx
